+++ assets/src/components/cd/pipelines/PipelineGraph.tsx
@@ -24,12 +24,13 @@ import chroma from 'chroma-js'
 import 'reactflow/dist/style.css'
 import styled, { useTheme } from 'styled-components'
 
+import { EdgeLineMarkerDefs, edgeTypes } from '../../utils/ReactFlowEdges'
+
 import { TestsNode } from './nodes/TestsNode'
 import { StageNode } from './nodes/StageNode'
 import { ApprovalNode } from './nodes/ApprovalNode'
 import { JobNode } from './nodes/JobNode'
 import { type DagreDirection, getLayoutedElements } from './utils/nodeLayouter'
-import { EdgeLineMarkerDefs, edgeTypes } from './EdgeLine'
 import { NodeType, getNodesAndEdges } from './utils/getNodesAndEdges'
 
 const nodeTypes = {
