+++ assets/src/components/cd/pipelines/utils/getNodesAndEdges.tsx
@@ -14,7 +14,8 @@ import { groupBy } from 'lodash'
 import { Entries } from 'type-fest'
 
 import { StageStatus, getStageStatus } from '../nodes/StageNode'
-import { PIPELINE_EDGE_NAME } from '../EdgeLine'
+
+import { MARKED_STEP_EDGE_NAME } from '../../../utils/ReactFlowEdges'
 
 import { reduceGateStates } from './reduceGateStatuses'
 import { DEMO_GATES, PIPELINE_DEBUG_MODE } from './_demo_data'
@@ -27,7 +28,7 @@ export enum NodeType {
 }
 
 const baseEdgeProps = {
-  type: PIPELINE_EDGE_NAME,
+  type: MARKED_STEP_EDGE_NAME,
   updatable: false,
 }
